WA-232 Refactor withdraw test
